move NumberField to new package
